+++ src/mkvtoolnix-gui/jobs/tool.h
@@ -29,9 +29,15 @@ protected:
   QAction *m_startAutomaticallyAction, *m_startManuallyAction, *m_viewOutputAction, *m_removeAction, *m_acknowledgeSelectedWarningsAction, *m_acknowledgeSelectedErrorsAction, *m_acknowledgeSelectedWarningsErrorsAction;
   QAction *m_openFolderAction, *m_editAndRemoveAction, *m_editCopyAction, *m_startImmediatelyAction;
   QMenu *m_jobQueueMenu, *m_jobsMenu;
+  QString m_searchTerm;
 
   mtx::gui::Util::FilesDragDropHandler m_filesDDHandler;
   QStringList m_droppedFiles;
+
+  enum SearchDirection {
+    Next,
+    Previous,
+  };
 
 public:
   explicit Tool(QWidget *parent, QMenu *jobQueueMenu);
@@ -68,6 +74,10 @@ public Q_SLOTS:
   void sortJobs(int logicalColumnIndex, Qt::SortOrder order);
   void hideSortIndicator();
 
+  void onFind();
+  void onFindNext();
+  void onFindPrevious();
+
   void onJobQueueMenu();
   void onContextMenu(QPoint pos);
   void moveJobsUpOrDown(bool up);
@@ -93,6 +103,10 @@ protected:
   virtual void dropEvent(QDropEvent *event) override;
 
   virtual void selectJobs(QList<Job *> const &jobs);
+
+  void querySearchTermFindAndSelectJob(SearchDirection const direction, int startRow);
+  void findAndSelectNextOrPreviousJob(SearchDirection const direction);
+  void findAndSelectJob(SearchDirection const direction, int startRow);
 };
 
 }